backend/requirements.txt
# ============================================================
# SkyMind Backend — Requirements (2026 Production)
# Amadeus API removed — using self-contained data pipeline
# ============================================================

# --- Core Framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# --- Database & ORM ---
supabase>=2.5.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9

# --- Auth & Security ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
bcrypt>=4.1.3

# --- HTTP & Async ---
httpx>=0.27.0
tenacity>=8.3.0
aiohttp>=3.9.5

# --- AI / ML ---
scikit-learn>=1.5.0
pandas>=2.2.2
numpy>=1.26.4
joblib>=1.4.2
xgboost>=2.0.3
pytz>=2024.1

# --- Payments ---
razorpay>=1.4.2

# --- SMS / Notifications ---
twilio>=9.2.2

# --- Scheduling ---
apscheduler>=3.10.4

# --- Pydantic V2 ---
pydantic>=2.7.0
pydantic-settings>=2.3.0
pydantic[email]>=2.7.0

# --- Fuzzy Search ---
rapidfuzz>=3.9.3

# --- Utilities ---
python-dateutil>=2.9.0
